Add microdata to <nav>
The default navigation list template 'nav_default.html5' of Contao is adding microdata to the list items (itemprop url and itemprop name).

Currently these Microdata inside the language links are hierarchically connected with the itemtype "WebPage" of the <body> element. This is wrong, because now all language links are interpreted as names and urls of the document itself. You can check this wrong connection and behaviour when testing any Contao site with installed ChangeLanguage in the microdata testing tool at https://search.google.com/structured-data/testing-tool

This PR is adding the correct itemtype "SiteNavigationElement" to the wrapping <nav> Element of the language links to correct the wrong connection to the body itemtype.